+++ src/AST.ml
@@ -91,6 +91,7 @@ type currency =
   | Vivid_crystallised_lifeforce
   | Primal_crystallised_lifeforce
   | Sacred_crystallised_lifeforce
+  | Fracturing_orb
   (* Not really currencies but... *)
   | Harvest_augment of Mod.harvest_tag
   | Harvest_non_to of Mod.harvest_tag
@@ -156,6 +157,7 @@ let show_currency = function
   | Vivid_crystallised_lifeforce -> "vivid_lifeforce"
   | Primal_crystallised_lifeforce -> "primal_lifeforce"
   | Sacred_crystallised_lifeforce -> "sacred_lifeforce"
+  | Fracturing_orb -> "fracture"
   | Harvest_augment tag -> "harvest_augment_" ^ Id.show (Mod.tag_id tag)
   | Harvest_non_to tag -> let t = Id.show (Mod.tag_id tag) in "harvest_non_" ^ t ^ "_to_" ^ t
   | Harvest_reforge tag -> "harvest_reforge_" ^ Id.show (Mod.tag_id tag)